+++ bcos-utilities/bcos-utilities/IOServicePool.h
@@ -32,83 +32,17 @@ public:
     using ExecutorType = boost::asio::io_context::executor_type;
     using Work = boost::asio::executor_work_guard<ExecutorType>;
     using WorkPtr = std::unique_ptr<Work>;
-    explicit IOServicePool(size_t _workerNum = std::thread::hardware_concurrency() + 1)
-      : m_works(_workerNum), m_nextIOService(0)
-    {
-        // create the ioservices
-        for (size_t i = 0; i < _workerNum; i++)
-        {
-            m_ioServices.emplace_back(std::make_shared<IOService>());
-        }
-    }
+    explicit IOServicePool(size_t _workerNum = std::thread::hardware_concurrency() + 1);
 
     IOServicePool(const IOServicePool&) = delete;
     IOServicePool& operator=(const IOServicePool&) = delete;
-    virtual ~IOServicePool() { stop(); }
+    virtual ~IOServicePool();
 
-    void start()
-    {
-        if (m_running)
-        {
-            return;
-        }
-        m_running = true;
-        for (size_t i = 0; i < m_ioServices.size(); ++i)
-        {
-            m_works[i] = std::make_unique<Work>(m_ioServices[i]->get_executor());
-        }
+    void start();
 
-        // one io_context per thread
-        for (const auto& ioService : m_ioServices)
-        {
-            // https://github.com/chriskohlhoff/asio/issues/932#issuecomment-968103444
-            m_threads.emplace_back([ioService, running = std::ref(m_running)]() {
-                if (!running)
-                {
-                    return;
-                }
+    std::shared_ptr<IOService> getIOService();
 
-                bcos::pthread_setThreadName("ioService");
-                ioService->run();
-            });
-        }
-    }
-
-    std::shared_ptr<IOService> getIOService()
-    {
-        auto selectedIoService = (m_nextIOService.fetch_add(1) % m_ioServices.size());
-        return m_ioServices.at(selectedIoService);
-    }
-
-    void stop()
-    {
-        if (!m_running)
-        {
-            return;
-        }
-        m_running = false;
-
-        // stop the work
-        for (auto& work : m_works)
-        {
-            work.reset();
-        }
-
-        // stop the io service
-        for (auto& ioService : m_ioServices)
-        {
-            ioService->stop();
-        }
-        // stop the thread
-        for (auto& thread : m_threads)
-        {
-            if (thread.joinable())
-            {
-                thread.join();
-            }
-        }
-        m_threads.clear();
-    }
+    void stop();
 
 private:
     std::vector<std::shared_ptr<IOService>> m_ioServices;
